# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vynel API",
    "description": "Vynel — desktop AI assistant for non-technical knowledge workers. Phase 1 ships locally; single-user.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/workspaces/{workspaceId}/channels": {
      "get": {
        "responses": {
          "200": {
            "description": "Array of Channel (without bot credentials)."
          },
          "404": {
            "description": "Workspace not found."
          }
        },
        "operationId": "getWorkspacesByWorkspaceIdChannels",
        "tags": [
          "channels"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          }
        ],
        "summary": "List connected channels for the workspace (owner-scoped; credentials excluded).",
        "x-sdk-name": "channels.list",
        "x-mcp": {
          "exposed": true,
          "name": "list_channels",
          "description": "List the connected messaging channels for the active workspace (owner-scoped). Returns each channel WITHOUT its bot credentials. Read-only."
        }
      }
    },
...
    "/workspaces/{workspaceId}/schedules": {
      "get": {
        "responses": {
          "200": {
            "description": "Array of Schedule."
          },
          "404": {
            "description": "Workspace not found."
          }
        },
        "operationId": "getWorkspacesByWorkspaceIdSchedules",
        "tags": [
          "schedules"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          }
        ],
        "summary": "List schedules for the active workspace (owner-scoped).",
        "x-sdk-name": "schedules.list",
        "x-mcp": {
          "exposed": true,
          "name": "list_schedules",
          "description": "List the scheduled routines for the active workspace (owner-scoped). Returns each schedule with its cron expression, destination, enabled flag, and next fire time."
        }
      },
      "post": {
        "responses": {
          "201": {
            "description": "Schedule created."
          },
          "400": {
            "description": "Invalid cron or missing channel."
          }
        },
        "operationId": "postWorkspacesByWorkspaceIdSchedules",
        "tags": [
          "schedules"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          }
        ],
        "summary": "Create a schedule (from a template or custom).",
        "x-sdk-name": "schedules.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "templateKind": {
                    "type": "string",
                    "enum": [
                      "morning-briefing",
                      "weekly-summary",
                      "email-watch",
                      "custom",
                      "reminder"
                    ]
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "cronExpression": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "timezone": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "promptTemplate": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 8000
                  },
                  "destinationKind": {
                    "type": "string",
                    "enum": [
                      "chat-only",
                      "chat-and-channel"
                    ]
                  },
                  "channelId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "catchUpOnMiss": {
                    "type": "boolean"
                  },
                  "approvalTimeoutMsOverride": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "fireAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "templateKind"
                ]
              }
            }
          }
        }
      }
    },
...
    "/channels": {
      "get": {
        "responses": {
          "200": {
            "description": "Array of Channel (without bot credentials)."
          }
        },
        "operationId": "getChannels",
        "tags": [
          "channels"
        ],
        "parameters": [],
        "summary": "List every channel the user owns — global + workspace (credentials excluded).",
        "x-sdk-name": "channelsUser.list",
        "x-mcp": {
          "exposed": true,
          "name": "list_my_channels",
          "description": "List every connected messaging channel the user owns — both global (no workspace) and workspace-scoped. Returns each channel WITHOUT its bot credentials. Read-only."
        }
      },
      "post": {
        "responses": {
          "201": {
            "description": "Channel connected (credentials excluded)."
          },
          "400": {
            "description": "Bot token invalid, unsupported kind, or workspaceId missing for a workspace scope."
          }
        },
        "operationId": "postChannels",
        "tags": [
          "channels"
        ],
        "parameters": [],
        "summary": "Connect a bot as a global or workspace channel (verifies the token before persisting).",
        "x-sdk-name": "channelsUser.connect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "scope": {
                        "type": "string",
                        "const": "global"
                      },
                      "channelKind": {
                        "type": "string",
                        "enum": [
                          "telegram",
                          "discord"
                        ]
                      },
                      "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "botCredentials": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "string"
                        }
                      },
                      "initialAllowedSenderId": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "scope",
                      "channelKind",
                      "displayName",
                      "botCredentials"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "scope": {
                        "type": "string",
                        "const": "workspace"
                      },
                      "workspaceId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "channelKind": {
                        "type": "string",
                        "enum": [
                          "telegram",
                          "discord"
                        ]
                      },
                      "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "botCredentials": {
                        "type": "object",
                        "additionalProperties": {
                          "type": "string"
                        }
                      },
                      "initialAllowedSenderId": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "scope",
                      "workspaceId",
                      "channelKind",
                      "displayName",
                      "botCredentials"
                    ]
                  }
                ]
              }
            }
          }
        }
      }
    },
...
    "/schedules": {
      "get": {
        "responses": {
          "200": {
            "description": "Array of Schedule."
          }
        },
        "operationId": "getSchedules",
        "tags": [
          "schedules"
        ],
        "parameters": [],
        "summary": "List every schedule the user owns — global + workspace.",
        "x-sdk-name": "schedulesUser.list",
        "x-mcp": {
          "exposed": true,
          "name": "list_my_schedules",
          "description": "List every scheduled routine the user owns — both global (no workspace) and workspace-scoped. Each has its cron expression (or one-time fire time), destination, enabled flag, and next fire time. Read-only."
        }
      },
      "post": {
        "responses": {
          "201": {
            "description": "Schedule created."
          },
          "400": {
            "description": "Invalid cron, missing channel, past fireAt, or workspaceId missing for a workspace scope."
          }
        },
        "operationId": "postSchedules",
        "tags": [
          "schedules"
        ],
        "parameters": [],
        "summary": "Create a global or workspace schedule (recurring cron OR one-time fireAt).",
        "x-sdk-name": "schedulesUser.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "scope": {
                        "type": "string",
                        "const": "global"
                      },
                      "templateKind": {
                        "type": "string",
                        "enum": [
                          "morning-briefing",
                          "weekly-summary",
                          "email-watch",
                          "custom",
                          "reminder"
                        ]
                      },
                      "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "cronExpression": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "timezone": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                      },
                      "promptTemplate": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 8000
                      },
                      "destinationKind": {
                        "type": "string",
                        "enum": [
                          "chat-only",
                          "chat-and-channel"
                        ]
                      },
                      "channelId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "catchUpOnMiss": {
                        "type": "boolean"
                      },
                      "approvalTimeoutMsOverride": {
                        "type": "integer",
                        "exclusiveMinimum": 0
                      },
                      "fireAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "scope",
                      "templateKind"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "scope": {
                        "type": "string",
                        "const": "workspace"
                      },
                      "workspaceId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "templateKind": {
                        "type": "string",
                        "enum": [
                          "morning-briefing",
                          "weekly-summary",
                          "email-watch",
                          "custom",
                          "reminder"
                        ]
                      },
                      "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "cronExpression": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "timezone": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                      },
                      "promptTemplate": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 8000
                      },
                      "destinationKind": {
                        "type": "string",
                        "enum": [
                          "chat-only",
                          "chat-and-channel"
                        ]
                      },
                      "channelId": {
                        "type": "string",
                        "minLength": 1
                      },
                      "catchUpOnMiss": {
                        "type": "boolean"
                      },
                      "approvalTimeoutMsOverride": {
                        "type": "integer",
                        "exclusiveMinimum": 0
                      },
                      "fireAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "scope",
                      "workspaceId",
                      "templateKind"
                    ]
                  }
                ]
              }
            }
          }
        }
      }
    },
...
  },
...
}